pyproject.toml
[tool.poetry]
name = "ts-v1"
version = "0.1.0"
description = ""
authors = ["josue-avila <josue.avila@olist.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^1.1.2"
SQLAlchemy = "^1.3.23"
python-dotenv = "^0.15.0"
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
psycopg2 = "^2.8.6"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
